+++ tailwind.config.js
@@ -5,7 +5,7 @@ const {
 
 /** @type {import('tailwindcss').Config} */
 module.exports = {
-  content: [
+  content: [ "*.html",
     "./src/pages/**/*.{js,ts,jsx,tsx,mdx}",
     "./src/components/**/*.{js,ts,jsx,tsx,mdx}",
     "./src/app/**/*.{js,ts,jsx,tsx,mdx}",
